+++ media_kit/lib/src/libmpv/core/native_library.dart
@@ -30,10 +30,10 @@ abstract class NativeLibrary {
         'libmpv.so',
       ],
       'macos': [
-        'libmpv.dylib',
+        'Mpv.framework/Mpv',
       ],
       'ios': [
-        'libmpv.dylib',
+        'Mpv.framework/Mpv',
       ],
     }[Platform.operatingSystem];
     if (names != null) {
@@ -49,9 +49,9 @@ abstract class NativeLibrary {
           'linux':
               'Cannot find libmpv at the usual places. Depending upon your distribution, you can install the libmpv package to make shared library available globally. On Debian or Ubuntu based systems, you can install it with: apt install libmpv-dev.',
           'macos':
-              'Cannot find libmpv.dylib. Check that it is present in the Frameworks folder of your application.',
+              'Cannot find Mpv.framework/Mpv. Check that it is present in the Frameworks folder of your application.',
           'ios':
-              'Cannot find libmpv.dylib. Check that it is present in the Frameworks folder of your application.',
+              'Cannot find Mpv.framework/Mpv. Check that it is present in the Frameworks folder of your application.',
         }[Platform.operatingSystem]!,
       );
     }