+++ media_kit_libs_ios_video/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.0.2
+
+- fix: use `mkdir` instead of `.gitkeep`
+
+## 1.0.1
+
+- fix: unable to publish iOS to AppStore
+- fix: support for iOS simulator
+- fix: add `.framework` & `.xcframework` for all libs
+
 ## 1.0.0
 
 - Initial release.