+++ media_kit_libs_linux/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+
+- Add `example/README.md`.
+
 ## 1.0.0
 
 - Initial release.