+++ media_kit_libs_windows_video/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.1
+
+- [mpv-player/mpv@`7990dd8`](https://github.com/mpv-player/mpv/commit/7990dd8f3f664e01c984a405dad7c8af320cabf1)
+- [ANGLE 2.1.18844 git hash: 2693b03eba82](https://github.com/google/angle)
+- MSVCRT
+- UCRT
+
 ## 1.0.0
 
 - [mpv-player/mpv@`7990dd8`](https://github.com/mpv-player/mpv/commit/7990dd8f3f664e01c984a405dad7c8af320cabf1)