+++ media_kit_video/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.4
+
+- fix: use `mkdir` instead of `.gitkeep`
+
+## 0.0.3
+
+- fix: add `.framework` & `.xcframework` for all libs
+
 ## 0.0.2
 
 - macOS support:
@@ -6,6 +14,8 @@
 - iOS support:
   - Hardware: MPV_RENDER_API_TYPE_OPENGL + pixel buffer
   - Software: MPV_RENDER_API_TYPE_SW + pixel buffer
+- fix(windows): use `TextureRegistrar::UnregisterTexture` release callback to free texture resources
+- fix(windows): synchronize texture unregister & release on frame dimensions change
 - feat: `aspectRatio` parameter for `Video` widget.
 
 ## 0.0.1